Ekle: dotnet code analyzer package, Düzelt: "SymFloat" signature correction, Düzelt: "Set" index base bug fix, Refactor: All models and most of the FrontService, Refactor: MatrisModel page, Refactor: Apply some CA18## and IDE fixes, Cleanup: Cleaned w/ cfg 2

+++ MatrisAritmetik.Core/Services/ISpecialMatricesService.cs
@@ -66,8 +66,8 @@ namespace MatrisAritmetik.Core.Services
         /// <param name="seed">Seed to use with <see cref="Random"/></param>
         /// <returns></returns>
         MatrisBase<dynamic> SymFloat(int dimension,
-                                     int min = 0,
-                                     int max = 1,
+                                     float min = 0,
+                                     float max = 1,
                                      dynamic seed = null);
 
         /// <summary>
